api package removed, interfaces splited between dao and service
