+++ src/main/java/r4v88/dao/AuthorDao.java
@@ -1,4 +1,4 @@
-package r4v88.api;
+package r4v88.dao;
 
 import r4v88.model.Author;
