+++ src/main/java/r4v88/dao/UserDaoImpl.java
@@ -1,6 +1,5 @@
 package r4v88.dao;
 
-import r4v88.api.UserDao;
 import r4v88.model.User;
 import r4v88.model.parser.UserParser;
